subrepo:
  subdir:   "third_party_modules/gtsam_module/gtsam"
  merged:   "abb912d"
upstream:
  origin:   "ssh://gerrit.skyd.io:29418/gtsam_upstream"
  branch:   "pull_on_6_14"
  commit:   "1ae7204"
git-subrepo:
  version:  "0.3.0"
  origin:   "https://github.com/ingydotnet/git-subrepo"
  commit:   "9a0f034"

+++ gtsam/nonlinear/NonlinearConjugateGradientOptimizer.h
@@ -142,7 +142,7 @@ boost::tuple<V, int> nonlinearConjugateGradient(const S &system,
 
   // GTSAM_CONCEPT_MANIFOLD_TYPE(V);
 
-  int iteration = 0;
+  size_t iteration = 0;
 
   // check if we're already close enough
   double currentError = system.error(initial);
